my_new_requirements.txt
## first....create the conda environment with python 3.10.0....
## either first install the requirement.txt and then otherrequirements.txt .... or directly install from this file but this will bring conflict on pdfminer.six (The user requested pdfminer.six==20221105
# pdfplumber 0.10.2 depends on pdfminer.six==20221105
# textract 1.6.5 depends on pdfminer.six==20191110)....
## dont install from this but can take the reference for all the versions for requiremnts.txt and other requirements.txt.. rather install first from requirements.txt and install otherrequirements.txt.... ok!!
accelerate==0.23.0
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
antlr4-python3-runtime==4.9.3
anyio==4.0.0
argcomplete==1.10.3
async-timeout==4.0.3
attrs==23.1.0
auto-gptq==0.2.2
backoff==2.2.1
beautifulsoup4==4.8.2
bitsandbytes-windows==0.37.5
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.0
chroma-hnswlib==0.7.2
chromadb==0.4.6
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
compressed-rtf==1.0.6
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.1
dataclasses-json==0.5.14
datasets==2.14.5
dill==0.3.7
diskcache==5.6.3
docx2txt==0.8
ebcdic==1.1.1
effdet==0.4.1
emoji==2.8.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
extract-msg==0.28.7
faiss-cpu==1.7.4
fastapi==0.99.1
filelock==3.12.4
filetype==1.2.0
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.37
greenlet==3.0.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
IMAPClient==2.1.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
InstructorEmbedding==1.0.1
iopath==0.1.10
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.267
langdetect==1.0.9
langsmith==0.0.43
layoutparser==0.3.4
llama-cpp-python==0.1.83
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.0
olefile==0.46
omegaconf==2.3.0
onnx==1.14.1
onnxruntime==1.15.1
openapi-schema-pydantic==1.2.4
opencv-python==4.8.1.78
openpyxl==3.1.2
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==9.5.0
portalocker==2.8.2
posthog==3.0.2
protobuf==3.20.2
psutil==5.9.5
pulsar-client==3.3.0
pyarrow==13.0.0
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.13
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
PyPDF2==3.0.1
pypdfium2==4.21.0
PyPika==0.48.9
pyreadline3==3.4.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.22
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
rapidfuzz==3.4.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rouge==1.0.1
rpds-py==0.10.6
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.12.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
SQLAlchemy==2.0.21
starlette==0.27.0
streamlit==1.27.2
streamlit-extras==0.0.9
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
textract==1.6.5
threadpoolctl==3.2.0
timm==0.9.7
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
transformers==4.34.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
unstructured==0.10.21
unstructured-inference==0.7.2
unstructured.pytesseract==0.3.12
urllib3==1.26.6
uvicorn==0.23.2
validators==0.22.0
watchdog==3.0.0
watchfiles==0.20.0
websockets==11.0.3
Werkzeug==3.0.0
xlrd==1.2.0
XlsxWriter==3.1.7
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
